Experiment - change architecture - All common Wrappers rewrited by deriving from PdfDictionary and PdfArray  - pros :     - remove circular reference in PdfObject <-> Wrapper, fix GC issue     - reduce object reference allocation in runtime     - globally less code and depth of invocation  - cons:     - no public api constraints to access raw data     - complex logic to detemine resulting type - reduce parsing speed     - ctors not allow other indirect objects read - affect fonts - Remove intermediate, redundant PdfDataObject move all logic to PdfDirectObject - Remove Wrap method with Activation.CreateInstance - Remove DictionaryWrapper - all dicts derive from PdfDictionary - Colorspaces - derive from PdfArray
Refactoring
- Globally remove unused usings
- Continue convert docs to c# native style

Breaking api
- Rename class PdfDocument to PdfCatalog
- Rename class PdfFile to PdfDocument
- Rename class Rectangle to PdfRectangle
- Rename property BaseObject to RefOrSelf
- Rename property BaseDataObject to DataObject
- Remove extension classes on single class use enums

+++ dotNET/PdfClown/Tokens/PdfDocEncoding.cs
@@ -24,7 +24,6 @@
 */
 
 using PdfClown.Util;
-using System;
 
 namespace PdfClown.Tokens
 {
@@ -80,7 +79,7 @@ namespace PdfClown.Tokens
             public override int GetKey(char value)
             {
                 return IsIdentity(value)
-                  ? (int)value
+                  ? value
                   : base.GetKey(value);
             }
